Update LICENSE
Existing Code Before the License Change:

Desktop4kids obtained the ProzillaOS code when it was still under the MIT License, Desktop4Kids is legally entitled to continue using and modifying that version of the code under the terms of the MIT License.

The MIT License is permissive and does not impose restrictions on future licensing changes by the original authors. Thus, Desktop4Kids can use the code it obtained under MIT without needing to adopt the new GPLv3 terms for that specific codebase.

New Changes or Updates Post-License Change:

Any updates or new code released after the change to GPLv3 would be subject to the terms of the GPLv3. If Desktop4Kids incorporate any of this new code into its project, the entire project would need to comply with GPLv3.
    
The GPLv3 license requires that any derivative works also be licensed under GPLv3, which means your modifications and the entire combined work would need to be open-sourced under the same license.

Combining Old and New Code:

If Desktop4Kids continue to use only the version of the code that was licensed under MIT, the modifications and derivative work do not need to comply with the new GPLv3 terms.

However, if Desktop4Kids combine its MIT-licensed code with any new code released under GPLv3, the combined work would need to be released under GPLv3.